fix: use consistent principal for collie based connectivity BB deploys
when running the terraform plan with the validation UAMI, validating
the "glaskugel" demo that does manual BB composition with collie
fails because it uses the azurerm provider's client_id as the owner.

This is convenient for executing the connectivity BB with meshStack
via the dedicated automation principal, but breaks the manual flow.

Implemented an explicit optional var + fallback to fix this.