feat: add support for automated tenant deletion via meshStack
Allow cancelling subscriptions in all landing zones of likvid-bank.
We can still control this setting individually from within meshStack
and e.g. force a manual approval flow. This is the most convenient way
to set this up for likvid bank for the time being.

We can later review if we e.g. want to exclude the permission for
landing zones hosting critical production workloads.